requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
Django==3.1
djangorestframework==3.12.0
idna==2.10
Markdown==3.2.2
pycparser==2.20
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
